+++ packages/widgetbook/test/src/addons/common/multi_addon_builder_test.dart
@@ -36,7 +36,7 @@ void main() {
           await tester.pumpWidgetWithMaterialApp(
             MultiAddonBuilder(
               addons: [
-                CustomAddOn(
+                CustomAddon(
                   initialSetting: 'test',
                 ),
               ],